+++ clang/lib/AST/ExprObjC.cpp
@@ -24,8 +24,9 @@
 using namespace clang;
 
 ObjCArrayLiteral::ObjCArrayLiteral(ArrayRef<Expr *> Elements, QualType T,
-                                   ObjCMethodDecl *Method, SourceRange SR)
-    : Expr(ObjCArrayLiteralClass, T, VK_PRValue, OK_Ordinary),
+                                   ObjCMethodDecl *Method, bool ECI,
+                                   SourceRange SR)
+    : ObjCObjectLiteral(ObjCArrayLiteralClass, T, ECI, VK_PRValue, OK_Ordinary),
       NumElements(Elements.size()), Range(SR), ArrayWithObjectsMethod(Method) {
   Expr **SaveElements = getElements();
   for (unsigned I = 0, N = Elements.size(); I != N; ++I)
@@ -37,22 +38,25 @@ ObjCArrayLiteral::ObjCArrayLiteral(ArrayRef<Expr *> Elements, QualType T,
 ObjCArrayLiteral *ObjCArrayLiteral::Create(const ASTContext &C,
                                            ArrayRef<Expr *> Elements,
                                            QualType T, ObjCMethodDecl *Method,
-                                           SourceRange SR) {
+                                           bool ECI, SourceRange SR) {
   void *Mem = C.Allocate(totalSizeToAlloc<Expr *>(Elements.size()));
-  return new (Mem) ObjCArrayLiteral(Elements, T, Method, SR);
+  return new (Mem) ObjCArrayLiteral(Elements, T, Method, ECI, SR);
 }
 
 ObjCArrayLiteral *ObjCArrayLiteral::CreateEmpty(const ASTContext &C,
                                                 unsigned NumElements) {
   void *Mem = C.Allocate(totalSizeToAlloc<Expr *>(NumElements));
-  return new (Mem) ObjCArrayLiteral(EmptyShell(), NumElements);
+  auto *ALE = new (Mem) ObjCArrayLiteral(EmptyShell(), NumElements);
+  ALE->setExpressibleAsConstantInitializer(NumElements == 0);
+  return ALE;
 }
 
 ObjCDictionaryLiteral::ObjCDictionaryLiteral(ArrayRef<ObjCDictionaryElement> VK,
                                              bool HasPackExpansions, QualType T,
-                                             ObjCMethodDecl *method,
+                                             ObjCMethodDecl *method, bool ECI,
                                              SourceRange SR)
-    : Expr(ObjCDictionaryLiteralClass, T, VK_PRValue, OK_Ordinary),
+    : ObjCObjectLiteral(ObjCDictionaryLiteralClass, T, ECI, VK_PRValue,
+                        OK_Ordinary),
       NumElements(VK.size()), HasPackExpansions(HasPackExpansions), Range(SR),
       DictWithObjectsMethod(method) {
   KeyValuePair *KeyValues = getTrailingObjects<KeyValuePair>();
@@ -72,14 +76,14 @@ ObjCDictionaryLiteral::ObjCDictionaryLiteral(ArrayRef<ObjCDictionaryElement> VK,
   setDependence(computeDependence(this));
 }
 
-ObjCDictionaryLiteral *
-ObjCDictionaryLiteral::Create(const ASTContext &C,
-                              ArrayRef<ObjCDictionaryElement> VK,
-                              bool HasPackExpansions, QualType T,
-                              ObjCMethodDecl *method, SourceRange SR) {
+ObjCDictionaryLiteral *ObjCDictionaryLiteral::Create(
+    const ASTContext &C, ArrayRef<ObjCDictionaryElement> VK,
+    bool HasPackExpansions, QualType T, ObjCMethodDecl *method, bool ECI,
+    SourceRange SR) {
   void *Mem = C.Allocate(totalSizeToAlloc<KeyValuePair, ExpansionData>(
       VK.size(), HasPackExpansions ? VK.size() : 0));
-  return new (Mem) ObjCDictionaryLiteral(VK, HasPackExpansions, T, method, SR);
+  return new (Mem)
+      ObjCDictionaryLiteral(VK, HasPackExpansions, T, method, ECI, SR);
 }
 
 ObjCDictionaryLiteral *
@@ -87,8 +91,10 @@ ObjCDictionaryLiteral::CreateEmpty(const ASTContext &C, unsigned NumElements,
                                    bool HasPackExpansions) {
   void *Mem = C.Allocate(totalSizeToAlloc<KeyValuePair, ExpansionData>(
       NumElements, HasPackExpansions ? NumElements : 0));
-  return new (Mem)
+  auto *DLE = new (Mem)
       ObjCDictionaryLiteral(EmptyShell(), NumElements, HasPackExpansions);
+  DLE->setExpressibleAsConstantInitializer(NumElements == 0);
+  return DLE;
 }
 
 QualType ObjCPropertyRefExpr::getReceiverType(const ASTContext &ctx) const {